requirements.txt
requests
psutil
scapy
requests
speedtest-cli
reportlab
requests


